Add parsing for multiple commands
We add split_commands, which takes user's input and splits it,
to form the individual commands from which the input consists of.
Moreover, we store the delimiters to a table, so it becomes
easier for the execution part to know how to combine the outputs
of successive commands. search_delimiter and add_delimiter both
help towards the above goals.