Add multiple pipe support
execute_command() was modified so multiple pipes can be detected
and executed. A single temp file wasn't adequate, because it can't
be used both as input and output for a command. As a result, two
seperate files were created, one for input and one for output
redirection. The contents of the output file are then copied back
to the input file, so they can serve as input for the next command.