Add cd support and minor tweaks
Support for cd command was added. It was implemented in parent,
because child is another environment, which results in no effect,
if executed there. Also modularized execute() with the addition
of input_redirect() and output_redirect(), for easier maintenance.
Lastly, check for status value was tweaked, to encompass the changes
happened, due to built-in function execution.